cloud_functions/snp_a005/requirements.txt
functions-framework==3.*
cloudevents==1.10.1
google-cloud-bigquery==3.20.1
google-cloud-storage==2.17.0
pandas==2.2.2
openpyxl==3.1.3
fsspec==2024.6.1
gcsfs==2024.6.1
db-dtypes==1.2.0
